Create proper vanilla R kernel image for use on k8s
Base the R kernel image on the pre-existing jupyter/r-notebook image.
This contains a vetted list of R-packages so all we add are the kernelspec
files, the bootstrap command and any (one) R-package necessary for the
kernel launcher.  This image will use the existing name: elyra/kubernetes-kernel-r

The previous image, based on the k8s-on-spark fork continues as before but
has been renamed to: elyra/kubernetes-spark-kernel-r

The R launch script changed the inclusion of the SparkR package from
using the library() function to using the require() since the latter does
not throw an exception if the package is not present.  This is more compatible
with spark-context-initialization-mode values of 'none' - used by the by
vanilla images.

Fixes #370