Add GIFs depicting scalability issues and solutions
Updated the main web page to use animated gifs to better depict the
"before" and "after" scenarios relative to deploying Enterprise Gateway.
This was done due to a comment I recieved at a conference from someone
that could not quickly and easily determine the purpose of EG, but having
seen the presentation, the benefits of EG made sense.  As a result, I
felt we could take those slides and incorporate them on the main page.

Unrelated: Our instructions for installing the IRkernel should include
the `curl` package as well.  Its absence was causing the
`devtools::install_github('IRkernel/IRkernel')` to be silently skipped
resulting missing master branch updates.