Configure use of namespaces for Enterprise Gateway and kernels
This moves the kubernetes-enterprise-gateway.yaml from the location in
the docker hierarchy to kubernetes and combines it with enterprise-gateway-acl.yaml.
The file is also renamed to enterprise-gateway.yaml.

The yaml is extended to create a primary namespace named `enterprise-gateway` in
which all EG-based objects are created.  The service account in the cluster role
binding has been updated to be the default account in the enterprise-gateway
namespace.

A namespace will be created for each requested kernel if the kernel creation
request does NOT provide a namespace name in the `KERNEL_NAMESPACE` environment
variable.  If provided, the namespace is assumed to exist.  If no KERNEL_NAMESPACE
value is provided, a namespace of the name <kernel-username>-<kernel-id> will be
created.  In addition, a role binding object will be created that references
the default service account for the new namespace.

When a namespace is created during kernel startup, EG will record that fact and
is responsible for deleting the namespace and all associated objects upon the
kernel's termination.

Currently, there is no means for running kernels within the same namespace as
enterprise-gateway unless KERNEL_NAMESPACE is set to 'enterprise-gateway'.  We
could easily add that option (such that setting the env is not required) at a
later time.

We will tighten up service accounts and role-based access control via #383

Fixes #381