Allow access from remote hosts
The update to Notebook 5.6.0 has code that prevents remote host access.  Since
the entire purpose of Enterprise Gateway is to allow remote host access, we need
to enable the web setting `allow_remote_access` to prevent the 403 exception.

Added a statement to update the jupyter framework files after finding issues
due to older dependencies.

Also increased the timeout for waiting for the docker image startup from 30
to 60 seconds and added the ability to override on the command line using
`PREP_TIMEOUT=n`.