Fix python_tf_gpu_kubernetes kernelspec
Fixed the image referenced in the kernel.json file to match that produced
by the build and in dockerhub.

Fixed the Makefile to include the python_tf_gpu_kubernetes directory in the
set of directories that require extension with kernel-pod.yaml and k8s
launch script.

Perform a second copy of source directories at end to allow per-kernel overrides
of files that are otherwise generic.  This would allow for, say, the python_tf_gpu
kernelspec directory to contain its own kernel-pod.yaml file that contains pod-
specific information.  Although not ideal, due to the redundancy it introduces
with the default kernel-pod.yaml file, it does allow for finer-grained customization
in a straightforward manner.

Fixes #425