Fix kernel launch failure detection
On python 3.x systems and reported in issues #432 and #435, the existing launch
failure detection was producing the following `TypeError` exception:

```
File "/opt/anaconda3/lib/python3.6/site-packages/enterprise_gateway/services/
             processproxies/processproxy.py", line 587, in detect_launch_failure
        if self.local_proc and self.local_proc.poll() > 0:
    TypeError: '>' not supported between instances of 'NoneType' and 'int'
```

This change checks if the result from `poll()` is None prior to checking its
value.

NOTE: This change should be added to the 1.x branch for a patch release.