Fix max-kernels-per-user when KERNEL_USERNAME not set
When the `KERNEL_USERNAME` is not provided in the kernel creation request,
the code that manages sessions would use a value of 'unspecified' while the
code that enforces limits would use the server's process username.  This
change uses a consolidated method so all locations render the same behavior
relative to defaulted values.

Note: These changes need to be merged into the 1.x branch where applicable.

Fixes #490