Add kernel-id parameter, cleanup launchers
Replaced the `{connection_file}` parameter in the `kernel.json` files with
`--RemoteProcessProxy.kernel-id {kernel_id}` since the former parameter was
used solely to extract the kernel's ID and create a new file to hold the
connection information.

The scala and python launchers will still honor `{connection_file}` but give
preference to the new parameter if both are provided.  They will also issue a
warning in such cases indicating that `{connection_file}` is deprecated.

The R launcher, however, uses the `argparser` library which does not support
optional positional arguments (like `{connection_file}` now becomes).  As a
result, the R launcher doesn't support the previous form.  This should only
be a problem where there `kernel.json` has not been updated, but the launcher
has been updated (which shouldn't happen).  I felt the trade-off between that
and switching to the `argparse` library was safer, since existing systems
won't necessarily have `argparse`.  That said, we may want to switch from
`argparser` to `argparse` at some point to acheive consistency with python.

I also enhanced the logging in the python and R launchers so poll requests do
not produce output (which happens every 3 seconds).

Fixed a dependency bug in `make kernelspecs` which was preventing the tar
files from getting rebuilt whenever one of their files was updated.

Fixes #500