Copy restart fix from 1.x (PR 514) to master
These changes are a copy of those used in PR #514...

When EG restarts remote kernels, it creates a new set of ports. This
was causing the activity_stream code, tied to iopub, to stop getting
updated. This change starts a new activity-watcher on restarts of
RemoteProcessProxy instances.

Added the capability to fetch a kernel's activity for use in test code.

Fixes #513