Package kernel launcher files for custom kernel images
This change reorganizes the kernel-launcher and kernel-bootstrap files that are
used when launching kernel images.  These files are then packaged into a tar.gz
file that can be downloaded from github relative to a given release.

As a result of this reorganization, the kernel-launchers sub-directory has been
moved to be co-located with the `bootstrap-kernel.sh` located in /usr/local/bin
by default.  This directory can be changed by specifying a different value for
KERNEL_LAUNCHERS_DIR when starting the kernel image.

The kernel image CMD entries have also been defaulted to the bootstrap-kernel.sh
file so kernel images can be launched w/o having to specify a command.  However,
attempts to manually launch these images will fail unless envs `KERNEL_ID` and
`EG_RESPONSE_ADDRESS` have been provided.

A _cutomization_ topic will be included as part of the efforts for #524.  This
will include how to create your own kernel image and corresponding kernelspec.

Note: I've also changed the `imagePullPolicy` for the enterprise-gateway image
from `Always` to `IfNotPresent` since it was updating to the wrong image that
I needed to test with prior to the PR.  This setting should be sufficient for
release-based images, but users should remove their dev based images if they
need to update to the latest dev tag.

Note 2: There's some odd issue happening with the kernel-r image.  The kernel-
spark-r image works fine.  Some kind of library issue.  I suspect there was a
change to the base notebook image that is now side-affecting ours.  Since Alan
is visiting the R images, I'm going to move on since the portions I've changed
test out.

Fixes #526