Update docs prior to 2.0.0 release
Made a pass to update various items and get them up to date in preparation
for the 2.0.0 release.  This also includes a section describing the necessary
items for building a custom kernel image - which is part of #528.  As a
result, it's probably best that this PR get merged after #528 to preserve
the integrity of its contents.

Fixes #524