Update files to be more PEP 8 compliant
Added docstrings to classes and methods and removed as many warnings as possible
to bring the core Enterprise Gateway code more in compliance with PEP 8.  Some
internal methods do not contain docstrings, but are commented.  This exercise
did not include updates to the launchers themselves.

As we make changes to files, we should do our best to adhere to PEP 8 and PEP 257
(regarding docstrings).

Fixes #407