Add support for run-as-UID/GID in containers
Added basic support for running as user id and group id values.  These values
can be specified via KERNEL_UID and KERNEL_GID respectively.  The defaults for
these values are 1000 (jovyan) and 100 (users) since that's how the containers
are built.  However, note that on k8s, use of `runAsGroup` requires that
feature-gate `RunAsGroup` first be enabled.  Failure to enable the feature-gate,
yet use either `runAsUser` or `runAsGroup` will result in the 0 (root) group
being the effective group of the container.  As a result, use of these attributes
are commented out in the `kernel-pod.yaml` file with an appropriate comment.
We also add the supplemental group of 100 (users) so that overrides of UID/GID
still permit the /home/jovyan directory to be writable.

The docker launcher has also been extended for these applicable values in docker.

Although Kubernetes PodSecurityPolicy objects can be used to ensure that the
root user/group is not used, a PSP affects ALL pods in the cluster, so we do
not want to introduce those in our yaml as that will break other applications.
UID and GID blacklists are sepecified via EG_{UID,GID}_BLACKLIST - both defaulting
to 0 (root).

Fixed up the env variable section of the docs tagging the sections as `text` so
various keywords are unexpectedly bolded. Also reformatted a recent change to
eliminate horizontal scroll bars.

Unrelated: Have noticed some of the restart tests have been failing because the
kernel status is returning 'starting' instead of 'idle' so we should tolerate
either.  Since another command is sent following restart, it will confirm the
kernel fully started.

Fixes #548