Fix k8s kernel restarts when EG creates kernel namespace
When EG creates the kernel pod's namespace (default behavior), restarts
would fail because the restart encouters an existing (but terminating)
namespace during the start portion of the restart (with the shutdown portion
having triggered the namespace's termination).  This change only deletes
the kernel pod during shutdowns associated with restarts, then tolerates
the namespace's existence after creation.  Note we still attempt the
namespace creation in case it was deleted outside of EG and a kernel
restart was attempted to rectify the situation.  This results in restart
behaviors identical to that if the user brought their own namespace.

Fixes #578