Tolerate errors when closing response-socket
While performing local debugging (on mac) I found that errno
57 "socket is not connecte" was being raised whenever EG went
to shutdown and close the response socket that it no longer needed.
Since the socket is no longer needed we should tolerate exceptions here.
This is probably a function of running the kernel locally as well as the
operating system (since this doesn't occur on RHEL systems).