Use a custom image as an init container to copy in kernelspecs.
This PR provides a way to mount kernelspecs into an Enterprise Gateway 
Kubernetes pod from a custom kernelspecs Docker image. This is an alternative 
to mounting kernelspecs into an Enterprise Gateway K8s pod via an NFS server. 
The motivation is two-fold:

Setting up, maintaining, and deploying kernelspecs to an NFS server is kind of 
a pain, especially within Kubernetes.

It'd be super nice for usability if a Helm "release" (Helm's terminology for a 
deployment) included everything necessary to get Enterprise Gateway up 
and running.. including kernelspecs. No multiple-step deploys. This works today 
with the stock kernelspecs (just deploy EG and go), but it'd also be nice if this 
worked for custom kernelspecs.

See #635 for more background.

Closes  #637