Add daemon to pull kernel images to each node
Kernel Image Puller (KIP) periodically polls Enterprise Gateway to
fetch the current set of configured kernelspecs, locating the kernel
image names from each.  It then pulls those image to the node on which
it is running.  Depending on its configured pull policy (KIP_PULL_POLICY),
it will either always attempt to pull each image (Always) or only pull
if it hasn't pulled that image before (IfNotPresent).

On Kubernetes, KIP is configured as a DaemonSet.  On Docker Swarm, it
will likely be configured as a global service (TBD).

Fixes #638