Add support for dynamic configurations
This change leverages a recent update to the traitlets library
that exposes an application's loaded configuration files.  When enabled
(via a positive value for `dynamic_config_interval`) it periodically
checks the loaded configuration files for updates (via file modification
times).  If updates are detected, the registered configurables are
updated with the contents of the files.  See documentation for more details.

Fixes #647