Update docker publish-related targets
Since we should build our published images from scratch (and not rely
on anything in the docker-cache), the following changes were made:

1. Added proper support to set the NO_CACHE macro.  Previously, this
was set to `--no-cache` but then not referenced in the docker build
rules. Instead, the default should be an empty value (`NO_CACHE?=`)
and have this "empty variable" referenced by the docker build target.
2. Changed the publish-* targets to push-* since all publish-* did was
push existing images.
3. Created a publish-images target that is responsible for performing
the full builds and pushes of all images. This target will set
NO_CACHE=--no-cache.
4. Unrelated. Found that sphinx-markdown-tables was preventing the
creation of new conda environments.  Since this is only for docs and
present in the docs environment.yml file, I commented it out of the
requirements.yml file at the project level.