Fix container termination signal handling in EG container
When pods are deleted, kubernetes will send a SIGTERM to the running process
to allow the process to gracefully terminate.  For Enterprise Gateway, this
should include the shutdown of each running kernel, including the appropriate
deletion of the kernel pod and, possibly, its namespace (if EG created one).
Because the ENTRYPOINT was a shell script, the signal was not propagating to
the python process. This change restores the already existing `tini` program
as the ENTRYPOINT (inherited from minimal-notebook) and sets the command
to `start-enterprise-gateway.sh` - which then execs the python script.  Once
in place, EG pod termination leads to the termination of running kernels as
well - just as is the case on bare metal installations.

Fixes #672