Refactor KernelSessionManager to accommodate various persistence methods (#645)
* Refactor KernelSessionManager to accommodate various persistence methods

This change does some minor refactoring of `KernelSessionManager` such
that it is now a base class with two abstract methods.  A new subclass,
`FileKernelSessionManager`, has been created to provide an example subclass
which writes to a file.  This will be the default implementation.

Other forms of persistence are encouraged, namely NoSQL or KV providers.

To use a different kernel session manager add the following to the EG
command line (or configuration file equivalent):

`--EnterpriseGatewayApp.kernel_session_manager_class=<MyKernelSessionManager>`

Note there may be some additional refactoring required once we determine
the needs for other kinds of persistence.

Fixes #644