Fix interrupts on intensive operations
When performing CPU-intensive operations within a cell, remote python
kernels can have their interrupt capabilities stop working due to
thread starvation.  This is because the gateway-listener thread is
unable to signal the kernel.  This change uses a `Process` rather
than a `Thread` in which to run the gateway-listener - which results
in it not being starved.  As a result, the signal can be sent to the
kernel and the cell's operation interrupted.

Fixes #682