Attempt to address build flakiness (#689)
Ran a series of different trials to try to figure this out.  There were two
primary issues.

1. nosetests were completing, but travis was not detecting that the nosetests
had completed.  So it never produced the green status that the test completed.
2. The build of the demo-base was encountering "github api rate limit" errors
when attempting to install SparkR.

Issue 1 appears to be (more frequently) resolved by using `travis_wait` with
a 3 minute timeout.  After 3 minutes it seems like travis breaks out of its
"coma" and sees that the the nosetests completed OK. Note: Detection of a
succesful nosetest in these "timeout" situations is still intermittent. I will
likely enlist the help of the travis team on this issue.

Issue 2 can be resolved by pushing demo-base to docker hub and NOT depending on
that image as part of the build for enterprise-gateway-demo.  This also trims
5-7 minutes from the overall travis time - aside from preventing the github rate
limit issue.

Since `sudo:` is deprecated, I've also removed it's setting.

Overall, build times have been trimmed by 6-8 minutes - and a bit more reliable.