Add helm chart tarball to distribution
Added more (hopefully useful) metadata to the Chart.yaml file that
includes making the version of the chart match that of the EG release.
Modified the release.sh to update the version information in the chart.
Added `helm-chart` target to Makefile and tied to `dist` target.
Updated docs to reference helm chart tarball on release page.

Fixes #656