Extend scala interrupt test timeout
The interrupt test timeout for scala was 30 seconds.  Since spark context
initialization can sometimes take longer (on systems with limited resources)
the test was failing - then causing subsequent failures.  Extending to 60
seconds resolves that issue.

Also changed how `VERSION` is set in the Makefile so it can be set outside
the Makefile in order to use different image versions, etc.

Closes #702