Remove dependency on Kernel Gateway
Copied the JupyterWebSocket, kernels, kernelspecs, sessions,
base handlers and mixins from Jupyter Kernel Gateway.  The
WebSocketPersonality was promoted as a top-level set of
handlers since the NotebookHttpPersonality was not copied
and there's no need to retain multiple personalities.

The following appear as authors on at least one of the files
copied from Juptyer Kernel Gateway...

Co-authored-by: Brian Burns <bburns@us.ibm.com>
Co-authored-by: Corey A. Stubbs <cstubbs@us.ibm.com>
Co-authored-by: Graham Dumpleton <Graham.Dumpleton@gmail.com>
Co-authored-by: Kevin Bates <kbates4@gmail.com>
Co-authored-by: Luciano Resende <lresende@apache.org>
Co-authored-by: Min RK <benjaminrk@gmail.com>
Co-authored-by: Nitin Dahyabhai <nitind@us.ibm.com>
Co-authored-by: Peter Parente <parente@gmail.com>
Co-authored-by: Roland Weber <rolweber@de.ibm.com>
Co-authored-by: SimonBiggs <mail@simonbiggs.net>