Removed dependency on Kernel Gateway (#734)
* Remove dependency on Kernel Gateway

Copied the JupyterWebSocket, kernels, kernelspecs, sessions,
base handlers and mixins from Jupyter Kernel Gateway.  The
WebSocketPersonality was promoted as a top-level set of
handlers since the NotebookHttpPersonality was not copied
and there's no need to retain multiple personalities.

* Fix lint errors and update version test

* Change KG_/kg_ to EG_/eg_ for env/settings references

For the environment variable default values, we'll go ahead and
reference the previous KG_ names only when the EG_ values are not
present. I.e., we leverage the default parameter with another
getenv call, but against the KG_ name, before returning the "fallback"
default.

* Update docs to reflect decoupling of JKG

* Use case-insensitive compare for boolean traitlet

The following appear as authors on at least one of the files
copied from Jupyter Kernel Gateway...

Co-authored-by: Brian Burns <bburns@us.ibm.com>
Co-authored-by: Corey A. Stubbs <cstubbs@us.ibm.com>
Co-authored-by: Graham Dumpleton <Graham.Dumpleton@gmail.com>
Co-authored-by: Kevin Bates <kbates4@gmail.com>
Co-authored-by: Luciano Resende <lresende@apache.org>
Co-authored-by: Min RK <benjaminrk@gmail.com>
Co-authored-by: Nitin Dahyabhai <nitind@us.ibm.com>
Co-authored-by: Peter Parente <parente@gmail.com>
Co-authored-by: Roland Weber <rolweber@de.ibm.com>
Co-authored-by: SimonBiggs <mail@simonbiggs.net>