Replace pycrypto with pycryptodomex
The pycrypto package has fallen out of maintenance, so this change
replaces it with pycryptdomex.  This package was chosen over pycryptodome
because it can co-exist with pycrypto, while pycryptodome cannot, and
we felt other applications may be using pycrypto.  If the remote kernel
launchers are not updated, the two are compatible since we use
the same mode.