Tolerate async updates coming on next notebook release (#811)
This commit supersedes the commit from PR #809.  The real issue
was that the `maybe_future` wrappers didn't account for the case
that they could be wrapping async/await-based methods and not
just gen.coroutines.  Updating to use the wrapper from the notebook
utilities allows EG to work with current and future notebook releases.

Resolves #806