Add username filtering for each kernel (#800)
Add support for filtering kernelspecs based on user and configured authorization.  These changes will remove a kernelspec from the resultset if `authorized_users` is configured and the user specified in the query parameter is not authorized to use the kernel.  If the kernelspec does not configure authorization the kernelspec will be included.