Pin jupyter-stack images to known working version
It was found that more recent base images (from jupyter stacks) are
breaking the spark-based images that derive from them.  It is suspected
that the move from conda 4.8.2 to 4.8.3 is side-effecting Spark in
some manner.  This change pins ALL jupyter-stack images to the same
tag (commit) for consistency even though some images are not used in
a Spark context.