Flow KERNEL_IMAGE value (#889)
It was recently discovered that KERNEL_IMAGE did not flow from
the client to be used during the kernel launch. Although this is a 
limited use case, as it requires the same image to be used for a specific
client regardless of the selected kernel, there are use cases where it
makes sense (see #884). This pull request makes the necessary changes
such that a payload env of KERNEL_IMAGE will be used for that request's
kernel launch.

The current definition of KERNEL_IMAGE is renamed to EG_KERNEL_IMAGE
and has no default. This env would be set if an administrator needed to use the
same image across all kernel launch requests (again, very limited utility). 
The precedence hierarchy, from lowest to highest, is as follows:

- Use EG_KERNEL_IMAGE.
- Use image_name from process_proxy.config from the 
   kernel.json file. (Common case)
- Use KERNEL_IMAGE from client.