Add docs and be stricter on bool variable.
In SSH tunneling section add information about GSS and Kerberos.
I realised that this is bool env, and that we likely want to be stricter
on which values are allowed, we don't want "0" to mean enable.

Thus only allow "True",  "False", empty string and unset (case
insensitive) as possible values.

Store the raw variable and its boolean interpretation separately for
error messages.