Validate images relative to python, spark, and debugger (#1030)
After creating and running images with the #986 and (then) #1029 PRs
while preparing for the EG 2.6.0 release, several issues were discovered,
mostly related to python, spark, and R versions. In addition, 
the tensorflow-gpu image was also using an older version of ipykernel.

One thing learned from this exercise is that PySpark on Spark 2.4.6
only supports Python 3.7 (or lower). We will need to revisit these base images
when moving to Spark 3.2 (in EG 3.0), but I suspect things will be a bit more 
straightforward and I'm hoping we can use synchronized tag values at that time.