Build images during release, capture release build/dist dirs (#1032)
It would be helpful to build the docker images during the release build. 
These will be pushed if the release is not in "dryRun", otherwise they 
remain on the local system. Previously, it was not sufficient to build the 
images from the release working directory because, after the release 
script has run, the dev versions are in place and extra care was required 
to get the proper kernelspecs included, etc.

The script also "captures" the build and dist trees to the 
version-specific directory. This will more easily enable troubleshooting 
and "adjustments" to build artifacts (if necessary).