fix(k8s): handle status of deleted Pod correctly
Whilst deleting a kernel namespace returns a v1.Status, deleting a Pod
will return a v1.Pod and its status field is a v1.PodStatus, not a
string.

The existing code currently raises an exception as it attempts to do an
'in' check against the V1PodStatus type rather than a string as it
expected. This is logged as:
> Error occurred deleting pod: argument of type 'V1PodStatus' is not iterable