Track kernels by tenant id
If tenant id is specified in the kernel start request's body via
tenant_id, or as a query parameter (?tenant_id) to GET /api/kernels,
it will be used to associate kernels with a tenant.  If no tenant ID is
given in the request body or as a query parameter the UNIVERSAL_TENANT_ID
will be used so that all managed kernels use the same logic.